update ID constraints, add UUID, small syntax refactor

+++ src/main/java/com/ikarabulut/shareable/repository/FileRepository.java
@@ -3,5 +3,5 @@ package com.ikarabulut.shareable.repository;
 import com.ikarabulut.shareable.models.FileModel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FileRepository extends CrudRepository<FileModel, Integer> {
+public interface FileRepository extends CrudRepository<FileModel, Long> {
 }
